Stored Procedure to get Best Value Games
This SP returns the games that have been given a high score by Metacritic and are relatively cheap (under 15$) and their popularity is gouged through the number of recommendations each game has received. Later this data is then matched with the player's table to get the number of hours played for each of these best-value games.